Added TimeLimited namespace
Includes three interfaes, partial implementation via three classes (two
are abstract), and one supporting enumeration.

IDateRange and its implementation provide a basic structure for
date/time ranges that enable or disable content.

ITimeLimited and its implementation provide a basic structure for
content types which utilize these date ranges, with a small number of
helper functions for determining when a piece of content is available.

ITimeLimitedCollection and its abstract implementation provide a basic
structure for holding a collection of a ITimeLimited content type, and
quickly getting the pieces of content which are currently available.