Rewrite of DateRange.OverlapsWith logic
Original code checked for date ranges that overlapped, and then assumed
that if none of those sets of conditions had been met, then the ranges
did not overlap.

New code goes the other way, checking that the ranges do not overlap,
and assuming they overlap if those conditions aren't met.